requirements.txt
wheel==0.34.2
networkx==2.4
pandas==1.0.4
matplotlib==3.2.1
gitpython==3.1.2
requests==2.23.0
pyyaml==5.3.1
progress==1.5
strsimpy==0.1.9
python-dateutil==2.8.1
sentistrength==0.0.8